+++ source/renderer/app/themes/daedalus/white.js
@@ -942,6 +942,8 @@ export const WHITE_THEME_OUTPUT = {
       'rgba(41, 181, 149, 0.05)',
     '--theme-transactions-filter-title-button-background-color-active':
       'rgba(41, 181, 149, 0.12)',
+    '--theme-transactions-transfer-funds-selected-wallet-background-color':
+      '#2d2d2d0d',
   },
   uploader: {
     '--theme-uploader-text-color': '#2d2d2d',