+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 ## vNext
 
+### Features
+
+- Implemented the wallet migration leftovers handling ([PR 2178](https://github.com/input-output-hk/daedalus/pull/2178))
+
 ### Chores
 
 - Dependency update ([PR 2190](https://github.com/input-output-hk/daedalus/pull/2190))